Moved file configuration classes into model module
