+++ cinnamon-platform/src/main/java/de/kiaim/cinnamon/platform/converter/StringToFileConfigurationConverter.java
@@ -1,7 +1,7 @@
 package de.kiaim.cinnamon.platform.converter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import de.kiaim.cinnamon.platform.model.file.FileConfiguration;
+import de.kiaim.cinnamon.model.configuration.data.file.FileConfiguration;
 import lombok.SneakyThrows;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.convert.converter.Converter;
