+++ cinnamon-platform/src/main/java/de/kiaim/cinnamon/platform/model/entity/FhirFileConfigurationEntity.java
@@ -1,14 +1,13 @@
 package de.kiaim.cinnamon.platform.model.entity;
 
 
-import de.kiaim.cinnamon.platform.model.file.FhirFileConfiguration;
-import de.kiaim.cinnamon.platform.model.file.FileType;
+import de.kiaim.cinnamon.model.configuration.data.file.FhirFileConfiguration;
+import de.kiaim.cinnamon.model.configuration.data.file.FileType;
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.Setter;
-import org.springframework.lang.Nullable;
 
 /**
  * File configuration for FHIR bundles.
