+++ cinnamon-platform/src/main/java/de/kiaim/cinnamon/platform/model/entity/XlsxFileConfigurationEntity.java
@@ -1,7 +1,7 @@
 package de.kiaim.cinnamon.platform.model.entity;
 
-import de.kiaim.cinnamon.platform.model.file.FileType;
-import de.kiaim.cinnamon.platform.model.file.XlsxFileConfiguration;
+import de.kiaim.cinnamon.model.configuration.data.file.FileType;
+import de.kiaim.cinnamon.model.configuration.data.file.XlsxFileConfiguration;
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
 import lombok.Getter;
