+++ cinnamon-platform/src/main/java/de/kiaim/cinnamon/platform/processor/CsvProcessor.java
@@ -15,9 +15,9 @@ import de.kiaim.cinnamon.platform.model.enumeration.CsvQuote;
 import de.kiaim.cinnamon.platform.model.enumeration.CsvRecordSeparator;
 import de.kiaim.cinnamon.platform.model.enumeration.DatatypeEstimationAlgorithm;
 import de.kiaim.cinnamon.platform.model.TransformationResult;
-import de.kiaim.cinnamon.platform.model.file.CsvFileConfiguration;
-import de.kiaim.cinnamon.platform.model.file.FileConfiguration;
-import de.kiaim.cinnamon.platform.model.file.FileType;
+import de.kiaim.cinnamon.model.configuration.data.file.CsvFileConfiguration;
+import de.kiaim.cinnamon.model.configuration.data.file.FileConfiguration;
+import de.kiaim.cinnamon.model.configuration.data.file.FileType;
 import de.kiaim.cinnamon.platform.service.DataSetService;
 import org.apache.commons.csv.CSVFormat;
 import org.apache.commons.csv.CSVPrinter;
