+++ cinnamon-platform/src/main/java/de/kiaim/cinnamon/platform/processor/DataProcessor.java
@@ -10,7 +10,7 @@ import de.kiaim.cinnamon.platform.model.dto.FileConfigurationEstimation;
 import de.kiaim.cinnamon.platform.model.entity.FileConfigurationEntity;
 import de.kiaim.cinnamon.platform.model.enumeration.DatatypeEstimationAlgorithm;
 import de.kiaim.cinnamon.platform.model.TransformationResult;
-import de.kiaim.cinnamon.platform.model.file.FileType;
+import de.kiaim.cinnamon.model.configuration.data.file.FileType;
 
 import java.io.InputStream;
 import java.io.OutputStream;
