+++ cinnamon-platform/src/main/java/de/kiaim/cinnamon/platform/processor/FhirProcessor.java
@@ -13,9 +13,9 @@ import de.kiaim.cinnamon.platform.model.entity.FhirFileConfigurationEntity;
 import de.kiaim.cinnamon.platform.model.entity.FileConfigurationEntity;
 import de.kiaim.cinnamon.platform.model.enumeration.DatatypeEstimationAlgorithm;
 import de.kiaim.cinnamon.platform.model.TransformationResult;
-import de.kiaim.cinnamon.platform.model.file.FhirFileConfiguration;
-import de.kiaim.cinnamon.platform.model.file.FileConfiguration;
-import de.kiaim.cinnamon.platform.model.file.FileType;
+import de.kiaim.cinnamon.model.configuration.data.file.FhirFileConfiguration;
+import de.kiaim.cinnamon.model.configuration.data.file.FileConfiguration;
+import de.kiaim.cinnamon.model.configuration.data.file.FileType;
 import de.unimuenster.imi.fhir.columns_parser.Column;
 import de.unimuenster.imi.fhir.transform.BundleTransformer;
 import de.unimuenster.imi.fhir.transform.ResourceExtractor;
